test: S12.21 promote SyslogFieldParser to Tests/Support
Lift the SYSLOG_FIELD_* indices, the SyslogField/SyslogMsg/SyslogMsgHasBom
parser, and the UTF8_BOM constant out of SolidSyslogTest.cpp into a shared
Tests/Support/SyslogFieldParser.{cpp,h} (it returns std::string, so C++).
Internal helpers (SkipSdata, FindFieldStart, SyslogMsgStart) stay file-static.
Standalone test refactor with no production change — the existing ~50 tests
prove the parser behaves identically; unblocks the S12.21 MessageFormatter
unit test (Slice 1).
